+++ apple-codesign/CHANGELOG.md
@@ -17,6 +17,9 @@ Released on ReleaseDate.
 * Added a hidden `debug-create-macho` command for generating Mach-O files.
   The command (and new code behind it) is intended to facilitate writing
   tests of Mach-O signing.
+* The `--code-signature-flags` argument of the `sign` command now correctly
+  applies multiple values. Before, flags were set to the final specified
+  value.
 * aws crates 0.53 -> 0.57.
 * bitflags 1.3 -> 2.0.
 * cryptographic-message-syntax 0.19 -> 0.25.